Started Level 2.
Added a level. There are currently rings the character has to travel on and destroy the enemies. Enemies drop time items which allow the character to get to the end of the level. Added materials for current and future levels. Added '=' key press to portal hub to go to the main menu. Set show mouse cursor to true for main menu. Updated README.md.