Started Level 3
Level 3 is a maze. Placed enemies and rival_boxes at strategic places, Objective is to reach portal at end of maze. There are dead ends. Placed windows in walls to see portal in one of the dead ends, but cannot jump thru windows to get to portal. There is also a lava hazard on correct path to portal. Final stretch is a gauntlet of enemies and rival boxes to reach portal.